test-support: promote the scenario generator to @satsuma/scenario-gen (sl-puky)
Feature 39's generated-input machinery lived in
satsuma-core/test/support/generated-scenarios.js: under test/, not compiled to
dist/, absent from core's exports map. Feature 41's lineage and graph properties
live in the CLI and the viz, so none of them could reach it.

Split the file along the seam it already had. The semantic model, the renderer,
the arbitraries and GENERATED_PROPERTY_PARAMETERS are pure string building and
move to a new private test-only package, tooling/satsuma-scenario-gen, whose only
dependency is fast-check and which has no build step. The four adapters that
import @satsuma/core stay behind in test/support/scenario-pipeline.js — renamed
because the file no longer generates anything.

The package must never depend on @satsuma/core: core's tests depend on it, so the
reverse edge would make core's test run need this package's output while this
package needed core's dist/. Exported typedefs gained a Scenario prefix, since
core's validate.ts already exports SemanticMapping/SemanticArrow/SemanticSchema
for the unrelated semantic-validation model.

Three core suites changed only their imports and all 679 core tests pass — three,
not the two sl-puky names: Feature 39 R4's coverage-oracle suite shipped after
the Feature 41 PRD recorded R4 as unticketed.

TypeScript consumers read the JSDoc types via allowJs + maxNodeModuleJsDepth: 1.
A narrower paths mapping to the source fails with TS6059 because the generator
sits outside satsuma-cli's rootDir, which tsc enforces even under noEmit; the
reasoning is recorded beside the options in tsconfig.test.json.
