+++ .tickets/sl-puky.md
@@ -1,6 +1,6 @@
 ---
 id: sl-puky
-status: open
+status: closed
 deps: []
 links: []
 created: 2026-08-03T16:23:38Z
@@ -22,3 +22,24 @@ New private workspace package tooling/satsuma-scenario-gen: type module, plain E
 
 satsuma-scenario-gen has no dependency on @satsuma/core and npm install from a clean checkout resolves without a cycle. Core's generated-coverage-properties.test.js and generated-format-properties.test.js keep every property they have and change only their import path; both pass. The CLI, viz and viz-backend test suites can import the package. Exported type names do not collide with core's validation model. Wired into npm run install:all and the repo checks.
 
+
+## Notes
+
+**2026-08-03T21:48:35Z**
+
+Cause: The Feature 39 generated-input machinery lived in
+satsuma-core/test/support/generated-scenarios.js — under test/, absent from
+core's exports map — so the CLI, viz and viz-backend suites that Feature 41's
+lineage properties need could not reach it.
+
+Fix: Promoted the pure half (model, renderer, arbitraries,
+GENERATED_PROPERTY_PARAMETERS) to a new private test-only package
+tooling/satsuma-scenario-gen with fast-check as its only dependency and no build
+step; core's pipeline adapters stay behind in test/support/scenario-pipeline.js.
+Exported typedefs gained the Scenario prefix. Three core suites (not two — the
+Feature 39 R4 coverage-oracle suite shipped after the PRD was written) changed
+only their import paths and all 679 core tests pass. TypeScript consumers read
+the JSDoc via allowJs + maxNodeModuleJsDepth: 1; the rejected paths-mapping
+alternative and its TS6059 failure are recorded in tsconfig.test.json and in
+features/41-lineage-graph-confidence/IMPLEMENTATION-NOTES.md. (commit
+immediately after 3ba4e95d)
