+++ tooling/satsuma-cli/package.json
@@ -2,7 +2,7 @@
   "name": "satsuma-cli",
   "version": "0.12.0",
   "private": true,
-  "description": "Satsuma CLI — LLM context slicer for Satsuma workspaces",
+  "description": "Satsuma CLI \u2014 LLM context slicer for Satsuma workspaces",
   "license": "MIT",
   "type": "module",
   "bin": {
@@ -32,6 +32,7 @@
     "web-tree-sitter": "^0.26.7"
   },
   "devDependencies": {
+    "@satsuma/scenario-gen": "file:../satsuma-scenario-gen",
     "@satsuma/viz-backend": "file:../satsuma-viz-backend",
     "@types/node": "^25.5.0",
     "c8": "^11.0.0",
