+++ tooling/satsuma-core/package.json
@@ -97,6 +97,7 @@
     "yaml": "^2.9.0"
   },
   "devDependencies": {
+    "@satsuma/scenario-gen": "file:../satsuma-scenario-gen",
     "@types/node": "^26.1.2",
     "c8": "^12.0.0",
     "fast-check": "^4.9.0",
