+++ tooling/satsuma-viz-backend/package.json
@@ -42,6 +42,7 @@
     "vscode-languageserver-types": "^3.18.0"
   },
   "devDependencies": {
+    "@satsuma/scenario-gen": "file:../satsuma-scenario-gen",
     "@types/node": "^25.5.0",
     "c8": "^11.0.0",
     "typescript": "^5.9.3"
