+++ tooling/satsuma-viz/package.json
@@ -21,6 +21,7 @@
   },
   "devDependencies": {
     "@lit-labs/ssr-dom-shim": "^1.6.0",
+    "@satsuma/scenario-gen": "file:../satsuma-scenario-gen",
     "@satsuma/viz-backend": "file:../satsuma-viz-backend",
     "@types/node": "^22.0.0",
     "c8": "^11.0.0",
